bias_models/requirements.txt
transformers>=4.41.1
torch>=2.2.2
sentencepiece>=0.2.0
langdetect>=1.0.9
python-dotenv>=1.0.1
tqdm>=4.66.4
numpy>=1.26.4
requests>=2.28.1
